bug fixes: Makefile, remove static declarations from gpio_ops.c, misc

+++ drivers/gpio/gpio_ioctl.h
@@ -13,8 +13,8 @@
 
 struct gpio_ioctl
 {
-    unsigned int pin;
-    unsigned int data;
+    uint8_t pin;
+    uint8_t data;
 };
 
 
@@ -26,8 +26,9 @@ struct gpio_ioctl
 #define SET_PIN _IOW(MAJOR_NUM, 3, struct gpio_ioctl *)
 
 /* The name of the device file */ 
-#define DEVICE_FILE_NAME "gpio_dev" 
-#define DEVICE_PATH "/dev/gpio_dev" 
+#define DEVICE_FILE_NAME "gpio_dev"
+#define DEVICE_PATH "/dev/gpio_dev"
 #define DEVFS_BUFMAX_SIZE 1024
 
 #endif
+
